+++ service/grails-app/domain/org/olf/erm/Entitlement.groovy
@@ -116,11 +116,15 @@ public class Entitlement implements MultiTenant<Entitlement> {
   ]
 
   static hasMany = [
-    coverage: HoldingsCoverage,
-    poLines: POLineProxy
+    coverage: HoldingsCoverage
   ]
 
   Set<HoldingsCoverage> coverage = []
+  
+  @OkapiLookup(
+    value = '/orders-storage/po-lines/${obj.poLineId}'
+  )
+  String poLineId
   
   static mappedBy = [
     coverage: 'entitlement',
@@ -160,6 +164,7 @@ public class Entitlement implements MultiTenant<Entitlement> {
              activeTo column: 'ent_active_to'
             authority column: 'ent_authority'
             reference column: 'ent_reference'
+             poLineId column: 'ent_po_line_id'
              coverage cascade: 'all-delete-orphan'
   }
 
@@ -217,6 +222,9 @@ public class Entitlement implements MultiTenant<Entitlement> {
               return val != null ?  ['externalEntitlement.reference.not.null'] : true
             }
           })
+         
+         
+         poLineId (nullable:true, blank:false)
   }
   
   @Transient