+++ service/grails-app/domain/org/olf/general/DocumentAttachment.groovy
@@ -3,8 +3,9 @@ package org.olf.general
 import grails.gorm.MultiTenant
 import com.k_int.web.toolkit.refdata.RefdataValue
 import com.k_int.web.toolkit.refdata.Defaults
+import org.olf.general.FileUpload
 
-class DocumentAttachment implements MultiTenant<DocumentAttachment> {
+class DocumentAttachment extends SingleFileAttachment implements MultiTenant<DocumentAttachment> {
 
   String id
   String name
@@ -14,7 +15,7 @@ class DocumentAttachment implements MultiTenant<DocumentAttachment> {
   Date dateCreated
   Date lastUpdated
 
-  @Defaults(['License', 'Misc', 'Consortium Negotiation Document'])
+  @Defaults(['License', 'Misc', 'Consortium negotiation document'])
   RefdataValue atType
 
   static mapping = {
@@ -37,7 +38,6 @@ class DocumentAttachment implements MultiTenant<DocumentAttachment> {
          atType(nullable:true, blank:false)
     dateCreated(nullable:true, blank:false)
     lastUpdated(nullable:true, blank:false)
-
   }
 
 }