+++ service/grails-app/migrations/module-tenant-changelog.groovy
@@ -5,4 +5,8 @@ databaseChangeLog = {
   include file: 'docs.groovy'
   include file: 'linked-licenses.groovy'
   include file: 'coverage-unification.groovy'
+  include file: 'update-mod-agreements-1-6.groovy'
+  include file: 'update-mod-agreements-1-7.groovy'
+  include file: 'suppl-docs.groovy'
+  include file: 'add-file-upload.groovy'
 }